requirements.txt
pyjwt==2.3.0
flask==2.0.2
python-dotenv==0.19.2
bcrypt==3.2.0
pytest==6.2.5
pytest-flask==1.2.0